+++ packages/ui/.storybook/stories-styles.css
@@ -9,6 +9,32 @@
 	h4,
 	h5,
 	h6 {
+		font-weight: 600;
 		border-bottom: none;
+	}
+}
+
+.colors {
+	display: flex;
+	width: 100%;
+}
+
+.color-item {
+	flex: 1;
+	display: flex;
+	flex-direction: column;
+	padding: 40px 10px 10px;
+	background: repeating-linear-gradient(
+		-55deg,
+		#d9d9d9,
+		#d9d9d9 1px,
+		transparent 1px,
+		transparent 6px
+	);
+	border: 1px solid #d9d9d9;
+
+	span {
+		font-size: 12px;
+		color: #000;
 	}
 }
