+++ packages/ui/package.json
@@ -17,7 +17,7 @@
 		"package:svelte": "svelte-kit sync && svelte-package",
 		"package:styles": "postcss ./src/styles/main.css -o ./dist/styles/main.css && pnpm run copy-fonts",
 		"copy-fonts": "postcss ./src/styles/fonts.css -o ./dist/fonts.css && cpy './src/styles/fonts/**/*.woff2' './dist/fonts' --parents",
-		"design-tokens:build": "npx co build && prettier --write ./src/lib/design-tokens.json ./src/styles/core/design-tokens.css",
+		"design-tokens:build": "npx tz build && prettier --write ./src/lib/design-tokens.json ./src/styles/core/design-tokens.css",
 		"prepublishOnly": "pnpm run package",
 		"prepare": "svelte-kit sync",
 		"storybook": "storybook dev --no-open -p 6006",
@@ -25,9 +25,6 @@
 	},
 	"devDependencies": {
 		"@chromatic-com/storybook": "^1.6.1",
-		"@cobalt-ui/cli": "^1.8.0",
-		"@cobalt-ui/plugin-css": "^1.7.3",
-		"@cobalt-ui/plugin-js": "^1.4.3",
 		"@csstools/postcss-bundler": "^1.0.15",
 		"@csstools/postcss-minify": "^1.1.5",
 		"@storybook/addon-docs": "^8.2.7",
@@ -42,6 +39,9 @@
 		"@sveltejs/kit": "catalog:svelte",
 		"@sveltejs/package": "^2.3.2",
 		"@sveltejs/vite-plugin-svelte": "catalog:svelte",
+		"@terrazzo/cli": "^0.0.11",
+		"@terrazzo/plugin-css": "^0.0.9",
+		"@terrazzo/plugin-js": "^0.0.1",
 		"autoprefixer": "^10.4.19",
 		"cpy-cli": "^5.0.0",
 		"postcss": "^8.4.38",
