+++ packages/ui/src/lib/design-tokens.json
@@ -3593,326 +3593,6 @@
 		}
 	},
 	"size": {
-		"2": {
-			"$type": "dimension",
-			"$value": "2px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:592",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"4": {
-			"$type": "dimension",
-			"$value": "4px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:594",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"6": {
-			"$type": "dimension",
-			"$value": "6px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:595",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"8": {
-			"$type": "dimension",
-			"$value": "8px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:596",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"10": {
-			"$type": "dimension",
-			"$value": "10px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:427:7275",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"12": {
-			"$type": "dimension",
-			"$value": "12px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:597",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"14": {
-			"$type": "dimension",
-			"$value": "14px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:2831:13998",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"16": {
-			"$type": "dimension",
-			"$value": "16px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:598",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"20": {
-			"$type": "dimension",
-			"$value": "20px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:599",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"24": {
-			"$type": "dimension",
-			"$value": "24px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:600",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"26": {
-			"$type": "dimension",
-			"$value": "26px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:2831:13999",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"28": {
-			"$type": "dimension",
-			"$value": "28px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:601",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"32": {
-			"$type": "dimension",
-			"$value": "32px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:602",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"36": {
-			"$type": "dimension",
-			"$value": "36px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:603",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"40": {
-			"$type": "dimension",
-			"$value": "40px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:604",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"48": {
-			"$type": "dimension",
-			"$value": "48px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:605",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"56": {
-			"$type": "dimension",
-			"$value": "56px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:606",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"64": {
-			"$type": "dimension",
-			"$value": "64px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:98:607",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"80": {
-			"$type": "dimension",
-			"$value": "80px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:2531:6326",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
-		"96": {
-			"$type": "dimension",
-			"$value": "96px",
-			"$description": "",
-			"$extensions": {
-				"mode": {},
-				"figma": {
-					"variableId": "VariableID:2531:6327",
-					"collection": {
-						"id": "VariableCollectionId:98:591",
-						"name": "size",
-						"defaultModeId": "98:0"
-					}
-				}
-			}
-		},
 		"icon": {
 			"$type": "dimension",
 			"$value": "16px",
@@ -3979,7 +3659,7 @@
 		},
 		"card-padding": {
 			"$type": "dimension",
-			"$value": "{size.14}",
+			"$value": "14px",
 			"$description": "",
 			"$extensions": {
 				"mode": {},
@@ -4086,7 +3766,7 @@
 			"useDTCGKeys": true,
 			"colorMode": "hex",
 			"variableCollections": ["clr-core", "clr", "size", "radius"],
-			"createdAt": "2024-07-28T12:37:45.122Z"
+			"createdAt": "2024-08-02T11:57:08.594Z"
 		}
 	}
 }
