+++ packages/ui/src/stories/b-design-tokens.mdx
@@ -16,7 +16,7 @@ For all basic styles like colors, typography, and spacing, radius, and shadows, 
 	height="600"
 	src="https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Fdesign%2FJir22AjzvPokdbMbe8yt5k%2F%25F0%259F%2592%258E-Core%3Fnode-id%3D0-1%26t%3DBvUKbga6mKnt6vo8-1"
 	allowFullScreen
-></iframe>
+/>
 
 ---
 
@@ -33,7 +33,13 @@ The first step here is to grab design tokens from Figma. To do this, you need to
 
 The second step is to convert the JSON file to a CSS file.
 
-You can generate a CSS file within this repository. Just replace the root file `design-tokens.json` with the new one and run the `build:design-tokens` command. The command will generate a new `design-tokens.css` file in the `packages/ui/src/styles/core` folder.
+You can generate a CSS file within this repository. Just replace the root file `design-tokens.json` with the new one and run:
+
+```bash
+$ pnpm design-tokens:build
+```
+
+The command will generate a new `design-tokens.css` file in the `packages/ui/src/styles/core` folder.
 
 In order to convert the JSON file to a CSS file, we use the [Cobalt UI CLI](https://cobalt-ui.pages.dev/guides/getting-started).
