+++ packages/ui/tsconfig.json
@@ -19,7 +19,7 @@
 		"./svelte-kit/types/**/$types.d.ts",
 		"vite.config.js",
 		"vite.config.ts",
-		"tokens.config.js",
+		"terrazzo.config.js",
 		"src/**/*.js",
 		"src/**/*.ts",
 		"src/**/*.svelte",
